Bugfix lang (#73)
* HISTORY.rst / changelog.txt : update changelog
* setup.py : update version number

* main.py: update how lang parameter is handled.
This fixes the incorrect handling of the --lang parameter.
Instead of changing the query in main.py, it is passed to
query.py where it modifies the url. 
Fixes https://github.com/taspinar/twitterscraper/issues/70
Fixes backward incompatability of query_tweets call in version 0.4.1

* update query.py
The new argument 'lang' is placed at the back to make it backward compatible.

* Update README.md with example code 
This commit updates the example code in the readme file. 
For windows it is necessary to declare "if __name__ == '__main__'" if multiprocessing is called in an imported module. 
This fixes https://github.com/taspinar/twitterscraper/issues/74